requirements.txt
# Core ML and data processing
torch==2.5.1+cu121
torchvision==0.20.1+cu121
torchaudio==2.5.1+cu121
numpy==2.1.2
pandas==2.3.0
scikit-learn==1.6.1
transformers==4.52.4
sentence-transformers==4.1.0

# Video and audio processing
opencv-python==4.10.0.84
moviepy==1.0.3
librosa==0.10.2
whisper-openai==1.0.0
ffmpeg-python==0.2.0

# LLM and embeddings
ollama==0.5.1
openai==1.85.0
langchain==0.3.25
langchain-core==0.3.64
langchain-community==0.3.24
langchain-ollama==0.3.3

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.41
pgvector==0.4.1
alembic==1.14.0

# Web framework
fastapi==0.115.9
uvicorn==0.34.3
pydantic==2.11.5
pydantic-settings==2.9.1
python-multipart==0.0.12

# Jupyter and development
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
jupyterthemes==0.20.0
theme-darcula==4.0.0


# Utilities
python-dotenv==1.1.0
requests==2.32.3
tqdm==4.67.1
pillow==11.0.0
typing-extensions==4.12.2
